Add ability to switch between real time and virtual time
This commit adds the ability to switch between real time and virtual time. Furthermore, it adds a test that shows the general idea. When switching from virtual time to real time, the remaining time of following events is delayed. A factor can be used to scale real time execution times.